정렬 버그 수정 중

+++ DXEngine/BitonicSort.h
@@ -5,7 +5,7 @@ namespace DE {
     class BitonicSort {
     public:
         struct Element {
-            uint32_t key;
+            float key;
             uint32_t value;
         };
